requirements.txt
asgiref==3.3.1
autopep8==1.6.0
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
idna==2.10
nose==1.3.7
pika==1.2.0
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
django-cors-headers>=3.2.1
coreapi>=2.3.3